Updated folder structure and removed non-jta-datasource
