--- finish/src/main/java/io/openliberty/guides/eventapp/resources/EventApplication.java
@@ -1,9 +1,0 @@
-package io.openliberty.guides.eventapp.resources;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("/event")
-public class EventApplication extends Application {
-
-}
