+++ finish/src/main/java/io/openliberty/guides/jpaguide/resources/EventApplication.java
@@ -1,0 +1,9 @@
+package io.openliberty.guides.jpaguide.resources;
+
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+@ApplicationPath("/event")
+public class EventApplication extends Application {
+
+}
