+++ finish/src/main/java/io/openliberty/guides/jpaguide/ui/util/ServiceUtil.java
@@ -1,4 +1,4 @@
-package io.openliberty.guides.eventapp.ui.util;
+package io.openliberty.guides.jpaguide.ui.util;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -12,7 +12,7 @@ import javax.ws.rs.client.Invocation.Builder;
 import javax.ws.rs.core.Form;
 import javax.ws.rs.core.Response;
 
-import io.openliberty.guides.eventapp.models.Event;
+import io.openliberty.guides.jpaguide.models.Event;
 
 public class ServiceUtil {
