+++ finish/src/main/liberty/config/server.xml
@@ -30,13 +30,5 @@
         <jdbcDriver libraryRef="derbyJDBCLib" />
         <properties.derby.embedded databaseName="JPAGuide" createDatabase="create"/>
     </dataSource>
-
-    <dataSource id="samplejpadatasourcenonjta"
-                jndiName="jdbc/samplejpadatasourcenonjta"
-                transactional="false">
-        <jdbcDriver libraryRef="derbyJDBCLib" />
-        <properties.derby.embedded databaseName="JPAGuide" createDatabase="create"/>
-    </dataSource>
-
-
+    
 </server>
